refactor: change Folder Structure
