+++ packages/gdsc-dju-admin/src/components/modal/ApplicantModal/styled.ts
