+++ packages/gdsc-dju-admin/src/pages/Email/index.tsx
@@ -5,7 +5,7 @@ import { useAtom } from 'jotai';
 import React, { useEffect, useState } from 'react';
 
 import EmailContainer from '../../components/organisms/EmailContainer';
-import AdminEmailCheckModal from '../../components/organisms/modal/AdminEmailCheckModal';
+import AdminEmailCheckModal from '../../components/modal/AdminEmailCheckModal';
 import SelectedEmailContainer from '../../components/organisms/SelectedEmailContainer';
 import { isDevelop } from '../../context/recruitInfo';
 import { db } from '../../firebase/firebase';
